refactoring : add comment
