--- Dataclass_Student.kt
@@ -1,13 +1,0 @@
-package csv
-
-data class Student(
-	val number: String = "",
-	val name:   String = "",
-	val kurasu: String = "",
-	val age:    Int    = 0,
-	val score:  Int    = 0)
-{
-	override fun toString(): String {
-        return "$number,$name,$kurasu,$age,$score"
-    }
-}
